algumas correções na view 5

+++ app/Http/Controllers/Secretary/Institution/InstitutionController.php
@@ -35,62 +35,57 @@ class InstitutionController extends Controller
 
     public function store(RegisterInstitutionRequest $request){
         $requestData = $request->validated();
-        
         $requestData['institution']['type'] = 'SCHOOL';
 
         DB::beginTransaction();
         try{
             $institution = Institution::create($requestData['institution']);         
             $institution->address()->create($requestData['address']);
-
             DB::commit();
-
-            return redirect()
-                ->route('secretary.institution.create')
-                ->with('success', 'Instituição cadastrada com sucesso.');
-
+            return $this->redirectBackWithSuccessAlert('Instituição criada com sucesso!');
         }catch(\Exception $exception){
             DB::rollBack();
-            return "Mensagem: " . $exception->getMessage();
+            return $this->redirectBackWithDangerAlert('Não foi possível criar a instituição!'.$exception->getMessage());
         };
     }
 
     public function show(Institution $institution){
         return view('secretary.institutions.show', [
             'institution' => $institution,
-            
         ]);
     }
 
     public function edit(Institution $institution)
     {   
-        $address = $institution->address;
-        
+        $address = $institution->address;   
         return view('secretary.institutions.edit', compact('institution', 'address'));
     }
 
-    public function update(Institution $institution, RegisterInstitutionRequest $request)
+    public function update($id, RegisterInstitutionRequest $request)
     {
-
+        $institution = Institution::findOrFail($id);
+        $address = Address::where('institution_id', $id)->firstOrFail();
         $requestData = $request->validated();
-        
-        $institution->update($requestData['institution']);
-        $address = $institution->address();
-        $address->update($requestData['address']);
 
-        return redirect()
-            ->route('secretary.institutions.index')
-            ->with('success', 'Instituição atualizada com sucesso!');
+        try{
+            $institution->update($requestData['institution']);
+            $address->update($requestData['address']);            
+            return redirect()
+                ->route('secretary.institution.index')
+                ->with('success', 'Instituição editada com sucesso!');
+        }catch(\Exception $exception){
+            return $this->redirectBackWithDangerAlert('Não foi possível editar a instituição!'.$exception->getMessage());
+        }
     }
 
     public function destroy(Institution $institution)
     {
-        
-        $institution->address()->delete();
-        $institution->delete();
-
-        return redirect()
-            ->route('secretary.institution.index')
-            ->with('success', 'Instituição deletada com sucesso!');
+        if($institution->users()->exists()){
+            return $this->redirectBackWithDangerAlert('Instituição ainda possui usuários.');;
+        }else{
+            $institution->address()->delete();
+            $institution->delete();
+            return $this->redirectBackWithSuccessAlert('Instituição deletada com sucesso!');
+        }
     }
 }
